fix(toast): fix toast with useSyncExternalStore to work toast methods propery
