+++ packages/toast/src/create-standalone-toast.tsx
@@ -7,16 +7,12 @@ import {
 } from "@chakra-ui/system"
 import defaultTheme from "@chakra-ui/theme"
 import * as React from "react"
-import {
-  CreateToastOptions,
-  ToastMethods,
-  ToastProvider,
-  ToastProviderProps,
-} from "./toast.provider"
-import { ToastId, ToastMessage } from "./toast.types"
+import { ToastProvider, ToastProviderProps } from "./toast.provider"
+import type { ToastId } from "./toast.types"
 import { UseToastOptions } from "./use-toast"
 import { createRenderToast } from "./toast"
 import { getToastPlacement } from "./toast.placement"
+import { toastStore } from "./toast.store"
 
 const defaults: UseToastOptions = {
   duration: 5000,
@@ -31,10 +27,6 @@ export interface CreateStandAloneToastParam
       }
     >,
     Omit<ToastProviderProps, "children"> {}
-
-export interface CallableToastMethods extends ToastMethods {
-  (message: ToastMessage, options?: CreateToastOptions): void
-}
 
 export const defaultStandaloneParam: CreateStandAloneToastParam &
   Required<Omit<CreateStandAloneToastParam, keyof ToastProviderProps>> = {
@@ -58,14 +50,11 @@ export function createStandaloneToast({
   toastSpacing,
   component,
 }: CreateStandAloneToastParam = defaultStandaloneParam) {
-  const ref = React.createRef<CallableToastMethods>()
-
   const colorModeContextValue = { colorMode, setColorMode, toggleColorMode }
   const ToastContainer = () => (
     <ThemeProvider theme={theme}>
       <ColorModeContext.Provider value={colorModeContextValue}>
         <ToastProvider
-          ref={ref}
           defaultOptions={defaultOptions}
           motionVariants={motionVariants}
           toastSpacing={toastSpacing}
@@ -87,28 +76,16 @@ export function createStandaloneToast({
   const toast = (options?: UseToastOptions) => {
     const normalizedToastOptions = normalizeToastOptions(options)
     const Message = createRenderToast(normalizedToastOptions)
-
-    return ref.current?.notify(Message, normalizedToastOptions)
+    return toastStore.notify(Message, normalizedToastOptions)
   }
-  /**
-   * Toasts can only be updated if they have a valid id
-   */
+
   toast.update = (id: ToastId, options: Omit<UseToastOptions, "id">) => {
-    if (!id) return
-
-    const normalizedToastOptions = normalizeToastOptions(options)
-    const Message = createRenderToast(normalizedToastOptions)
-
-    ref.current?.update(id, {
-      ...normalizedToastOptions,
-      message: Message,
-    })
+    toastStore.update(id, normalizeToastOptions(options))
   }
 
-  toast.notify = ref.current?.notify ?? noop
-  toast.closeAll = ref.current?.closeAll ?? noop
-  toast.close = ref.current?.close ?? noop
-  toast.isActive = ref.current?.isActive ?? noop
+  toast.closeAll = toastStore.closeAll
+  toast.close = toastStore.close
+  toast.isActive = toastStore.isActive
 
   return {
     ToastContainer,
