+++ packages/toast/src/toast.utils.ts
@@ -59,3 +59,39 @@ export function getToastStyle(position: ToastPosition): React.CSSProperties {
     alignItems,
   }
 }
+
+/**
+ * Compute the style of a toast based on its position
+ */
+export function getToastListStyle(
+  position: ToastPosition,
+): React.CSSProperties {
+  const isTopOrBottom = position === "top" || position === "bottom"
+  const margin = isTopOrBottom ? "0 auto" : undefined
+
+  const top = position.includes("top")
+    ? "env(safe-area-inset-top, 0px)"
+    : undefined
+  const bottom = position.includes("bottom")
+    ? "env(safe-area-inset-bottom, 0px)"
+    : undefined
+  const right = !position.includes("left")
+    ? "env(safe-area-inset-right, 0px)"
+    : undefined
+  const left = !position.includes("right")
+    ? "env(safe-area-inset-left, 0px)"
+    : undefined
+
+  return {
+    position: "fixed",
+    zIndex: 5500,
+    pointerEvents: "none",
+    display: "flex",
+    flexDirection: "column",
+    margin,
+    top,
+    bottom,
+    right,
+    left,
+  }
+}
