+++ packages/toast/stories/toast.stories.tsx
@@ -420,4 +420,4 @@ export const ToastWithCustomIcon = () => {
       <Button onClick={() => toast.close(id)}>Close One</Button>
     </ButtonGroup>
   )
-} 
+}
